Separate test framework from tests themselves
