+++ tests/framework/evaluators/abstract_agent_evaluator.py
@@ -17,8 +17,8 @@ from leaf_common.parsers.dictionary_extractor import DictionaryExtractor
 
 from neuro_san.message_processing.basic_message_processor import BasicMessageProcessor
 
-from tests.neuro_san.client.agent_evaluator import AgentEvaluator
-from tests.neuro_san.client.assert_forwarder import AssertForwarder
+from tests.framework.interfaces.agent_evaluator import AgentEvaluator
+from tests.framework.interfaces.assert_forwarder import AssertForwarder
 
 
 class AbstractAgentEvaluator(AgentEvaluator):
